Move hikari transactor creation definition and implementation to `storage`
